1st commit of adding enable/disable and over write config.yaml. (#10)
* 1st commit of adding enable/disable and over write config.yaml.

* 2nd commit of adding enable/disable menu and checking if menu was installed before.

* 3rd commit of adding enable/disable menu and checking if menu was installed before.

+++ install.sh
@@ -31,22 +31,33 @@ else
 fi
 
 # Install ROS menu and config file
+if [ -f ~/ros_menu/config.yaml ]; then
+    echo  "The Neuron Startup Menu was already installed!"
+    echo -n "Do you want to reinstall? (Your setting in the file config.yaml would be over written.) (y/N): "
+    read over_write
+else
+    over_write = "y"
+fi
+if ! [ "$over_write" '==' "y" ] || [ "$over_write" '==' "Y" ]; then
+    echo "Skip installing Neuron Startup Menu!"
+    exit
+fi
 rm -f ~/.ros_menu
 ln -s `pwd` ~/.ros_menu
 if [[ -n $1 ]]; then
     config_file=./yaml/$1
 fi
 cp $config_file config.yaml
-
 if ! grep -q ros_menu ~/.${shell}rc; then
     cat <<EOF >> ~/.${shell}rc
 # Neuron Startup Menu #
 ros_bashrc_path=~/.ros_menu/ros_bashrc
 if [ -f \$ros_bashrc_path ]; then
-    source \$ros_bashrc_path
+   source \$ros_bashrc_path
 fi
 # End of Neuron Startup Menu #
 EOF
 fi
+
 
 echo "Neuron Startup Menu installed successfully"
